Paired with Lachlan
We spent most of our time thinking about what structure to use in order to store
the grid and cell-state. We settled on using a Set that would contain Cell
objects. A Cell object would be identified by its coordinates, and any two Cell
objects with the same coordinates would be considered equal. In this way, we
guarantee that no two cells in the Set would have the same coordinates. Each
cell would then be told to transform on each generation, and it would be able to
find the state of each adjacent cell by performing the coordinate math based on
its own coordinates.